refactor: テーブルエディタのリファクタリング
